加入缓存机制

+++ e3parent/e3-manager/e3-manager-dao/src/main/java/com/kk/mapper/ContentMapper.java
@@ -20,7 +20,7 @@ public interface ContentMapper {
 
     List<Content> selectByExample(ContentExample example);
 
-    Content selectByPrimaryKey(Long id);
+    Content selectByPrimaryKey(int id);
 
     int updateByExampleSelective(@Param("record") Content record, @Param("example") ContentExample example);
